+++ src/WDCMethods.js
@@ -1,6 +1,8 @@
 import { get } from "./utils.js";
 import { locationMode } from "./enums.js";
 import { notify } from "./notifications.js";
+import { parse, toSeconds } from "iso8601-duration";
+var format = require("date-format");
 
 /*global  tableau:true*/
 
@@ -94,6 +96,18 @@ const formatJSONAsTable = (data, tableName) => {
 };
 
 /*
+generates a query-ready ISO 8601 with timezone date time from a datetime and timezone string.
+*/
+const generateDateTime = (timeZone, dateTime, queryMode) => {
+  if (queryMode) {
+    // whether or not we need to escape '+'
+    return `${dateTime.substring(0, 16)}${timeZone.replace("+", "%2b")}`;
+  } else {
+    return `${dateTime.substring(0, 16)}:00.000${timeZone}`;
+  }
+};
+
+/*
 generates a URL for query parameters contained in the connectionData object accepted as an argument
 */
 const generateURL = connectionData => {
@@ -102,6 +116,10 @@ const generateURL = connectionData => {
   let locationQuery = "";
   let siteTypeQuery = "";
   let agencyCodeQuery = "";
+  let durationCodeQuery = "";
+  let modifiedSinceCodeQuery = "";
+  let temporalRangeQuery = "";
+  let historical = "";
 
   switch (connectionData.locationMode) {
     case locationMode.SITE: {
@@ -114,7 +132,6 @@ const generateURL = connectionData => {
       break;
     }
     case locationMode.COORDS: {
-      // west south east north
       let bounds = connectionData.boundaryCoords;
       locationQuery = `&bBox=${bounds.west},${bounds.south},${bounds.east},${bounds.north}`;
       break;
@@ -140,7 +157,60 @@ const generateURL = connectionData => {
     agencyCodeQuery = `&agencyCd=${connectionData.agencyCode}`;
   }
 
-  return `https://waterservices.usgs.gov/nwis/iv/?format=json${locationQuery}&period=P1D${paramQuery}${siteTypeQuery}${agencyCodeQuery}&siteStatus=all`;
+  if (connectionData.durationCodeActive) {
+    durationCodeQuery = `&period=${connectionData.durationCode}`;
+
+    let periodHistorical =
+      parseInt(toSeconds(parse(connectionData.durationCode))) >= 10368000; //approximate 120 days in seconds
+    if (periodHistorical) {
+      historical = "nwis.";
+    }
+  }
+
+  if (connectionData.temporalRangeActive) {
+    let startDateString = generateDateTime(
+      connectionData.temporalRangeData.timeZone,
+      connectionData.temporalRangeData.startDateTime,
+      true
+    );
+    let endDateString = generateDateTime(
+      connectionData.temporalRangeData.timeZone,
+      connectionData.temporalRangeData.endDateTime,
+      true
+    );
+    temporalRangeQuery = `&startDT=${startDateString}&endDT=${endDateString}`;
+
+    if (typeof connectionData.currentDateTime === "string") {
+      // this is necessary because JSON.stringify/JSON.parse are not symmetrical with respect to Date objects
+      // JSON.stringify converts date objects to strings, so they must be manually reconstructed as Date objects
+      // we do this with a formatting library, as behavior of Date() for parsing format strings is not standardized in older browsers.
+      connectionData.currentDateTime = format.parse(
+        format.ISO8601_WITH_TZ_OFFSET_FORMAT,
+        connectionData.currentDateTime
+      );
+    }
+    let startDate = format.parse(
+      format.ISO8601_WITH_TZ_OFFSET_FORMAT,
+      generateDateTime(
+        connectionData.temporalRangeData.timeZone,
+        connectionData.temporalRangeData.startDateTime,
+        false
+      )
+    );
+    if (
+      connectionData.currentDateTime.getTime() - startDate.getTime() >=
+      10368000000
+    ) {
+      //approximate 120 days in milliseconds
+      historical = "nwis.";
+    }
+  }
+
+  if (connectionData.modifiedSinceCodeActive) {
+    modifiedSinceCodeQuery = `&modifiedSince=${connectionData.modifiedSinceCode}`;
+  }
+
+  return `https://${historical}waterservices.usgs.gov/nwis/iv/?format=json${locationQuery}${paramQuery}${siteTypeQuery}${agencyCodeQuery}${durationCodeQuery}${modifiedSinceCodeQuery}${temporalRangeQuery}&siteStatus=all`;
 };
 
 /*
@@ -280,5 +350,6 @@ export {
   generateSchemaTablesFromData,
   getTimeSeriesByID,
   reformatTimeString,
-  sanitizeVariableName
+  sanitizeVariableName,
+  generateDateTime
 };