+++ CHANGELOG.md
@@ -29,5 +29,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added options for Test and QA buckets as targets for Jenkins builds
 ### Changed
 - Changed styling so all elements are left justified
-
+- Added various temporal query parameters